pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "proton-cleanup"
version = "0.2.10"
description = "Scan and clean up Steam Proton compatibility environments"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Debashis Dip", email = "nihan.dip@gmail.com" },
]
license-files = ["LICENSE"]
requires-python = ">=3.10"
keywords = ["steam", "proton", "wine", "steam-deck", "cleanup"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "vdf>=3.4",
    "PySide6-Essentials>=6.7",
]

[project.urls]
Homepage = "https://github.com/d1p/proton-cleanup"
Repository = "https://github.com/d1p/proton-cleanup"
Issues = "https://github.com/d1p/proton-cleanup/issues"

[project.scripts]
proton-cleanup = "proton_manager.cli:main"
proton-manager = "proton_manager.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

# ---------------------------------------------------------------------------
# Hatch build configuration
# ---------------------------------------------------------------------------

[tool.hatch.build.targets.wheel]
packages = ["src/proton_manager"]

# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
